[vm] Support loading unboxed float consts to stack
Split off https://dart-review.googlesource.com/c/sdk/+/284300.

Doubles are always loaded to DoubleStackSlots, and should already
be covered.

TEST=tested in dependent CL.
If the code was moved to the assemblers, and the assemblers were given
access to `Location`, this could be unit tested in an assembler test.
See issue for discussion.

Closes: https://github.com/dart-lang/sdk/issues/53900
Change-Id: Iaed2cf1fadebebd1857709cb5d7cc3b41a3d9747
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/332761
Reviewed-by: Ryan Macnak <rmacnak@google.com>
Commit-Queue: Daco Harkes <dacoharkes@google.com>