[infra/config] Increase flutter-analyze builder timeout to 60 minutes
The builder is consistently taking around 45 minutes, which is the
current timeout. It is starting to fail due to this.

The timeout was increased from 25 minutes to 45 minutes in
https://dart-review.googlesource.com/c/sdk/+/201680

The timeout is not increased to the default 3 hours because
it is important to know when the time taken to analyze
Flutter source increases significantly.

Change-Id: Ic927610769819561d4726ab42a8fbe4418825dec
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/332962
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Commit-Queue: William Hesse <whesse@google.com>
Reviewed-by: Devon Carew <devoncarew@google.com>