[infra] allow to specify arch twice.
A recent change: https://dart-review.googlesource.com/c/sdk/+/333842
fixed a check in the test runner. The check fails when a flag
is provided to test.py but it is also implied by the configuration
name.

However, The web macos bots currently need to specify the arch flag,
even though it is implied by the named configuration. Without doing so
the sharding infrastructure seems to fail to assign the correct
architecture for the spawn bots.

This CL temporarily allows to continue specifying this flag to keep the
bots running. Ideally we should follow up with a infrastructure fix
to ensure this flag is not needed in the future.

This is being tracked in https://github.com/dart-lang/sdk/issues/53948

Change-Id: Ifba6a8924cbca9f3aca38933113abb168ff405d4
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/333705
Reviewed-by: Nicholas Shahan <nshahan@google.com>
Commit-Queue: Sigmund Cherem <sigmund@google.com>