Switch from leak_tracker/leak_tracker to leak_tracker/memory_usage.
It is relanding of https://dart-review.googlesource.com/c/sdk/+/334143

Contributes to https://github.com/flutter/flutter/issues/135856

We want Flutter to take dependency on leak_tracker. It will be easier to manage in G3 if Flutter and Dart SDK depend on different packages. So, separated memory_usage and switching analyzer to it.

Change-Id: Ic85172bfe28cf070e01f5ff298ee975b25ae6f68
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/335061
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Oleh Prypin <oprypin@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Polina Cherkasova <polinach@google.com>