[infra] increase timeout for loading the sdk in ddc.
This increases the timeout from 30 to 45s when using require.js.
We noticed that many tests (>50) flake on firefox due to this timeout,
so we hope that adding a bit of extra time may help reduce the
likelyhood of tests flaking for this reason.

Bug: https://github.com/dart-lang/sdk/issues/53991

Change-Id: I3dd952d5dd42c8644f96a636df9a7cd486312ad6
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/335503
Auto-Submit: Sigmund Cherem <sigmund@google.com>
Commit-Queue: Nicholas Shahan <nshahan@google.com>
Reviewed-by: Nicholas Shahan <nshahan@google.com>