[tool] Improve bisection script
The commit ranges reported are inclusive the oldest commit. This is
signified by the `~` in the range `xxx~...xxx`.

This CL fixes the script to take this oldest commit into account. It
used to be ignored by the `git log` call omitting the `~`.
Bug: bisection on https://github.com/dart-lang/sdk/issues/54029

This CL also fixes the range prints to be GitHub paste-able. GitHub
doesn't auto syntax highlight `xxx~...xxx`. So instead look up the
parent commit of the oldest commit and use `xxx...xxx` to print the
commit range.

Finally, this CL enables swapping the hashes in the range in the
command-line arguments.

Change-Id: Ic5c685fee2941e7b50c221770962edebe55e62f2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/335801
Reviewed-by: Martin Kustermann <kustermann@google.com>
Commit-Queue: Daco Harkes <dacoharkes@google.com>