Add an AST-to-IR conversion stage to the Wolf analysis prototype.
This CL continues pulling a steel thread through the design of Wolf
analysis, by adding a stage that converts from analyzer ASTs to the
Wolf analysis IR. The set of supported instructions is still quite
minimal, so pretty much the only functions that can be handled take
the form `functionName() => literalValue;`. I plan to start adding
more instructions next.

A simple IR interpreter is included, solely for use in unit tests, so
that in the unit tests for AST-to-IR conversion, we can actually
verify that the IR properly reflects the intended semantics.

Includes a minor bug fix to the validator and improved validator
testing.

Also, a README.md file is added, giving an overview of the design of
Wolf analysis and noting that the code as currently experimental.

Change-Id: I1df724a9987c9eed205591cf5f3b8a3ec59cb864
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/335402
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Phil Quitslund <pquitslund@google.com>