[package:vm_service] Fix flakiness of uri_mappings_lookup_test
This test has been failing flakily on Windows because sometimes one of
the resolved URIs contains an uppercase drive letter and sometimes it
contains a lowercase one. This CL fixes this problem by doing a
case-insensitive check of that resolved URI on Windows, as paths on
Windows are case-insensitive.

TEST=CI

Fixes: https://github.com/dart-lang/sdk/issues/54034
Change-Id: I67379930baca938da12efe53d05f05284f71e325
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/335920
Reviewed-by: Ben Konyi <bkonyi@google.com>
Commit-Queue: Derek Xu <derekx@google.com>